** NOTE ** VERSIONS MOVING FORWARD WILL NO LONGER HAVE LANDSCAPE ORIENTATION SUPPORT! ** NOTE ** Task detail now includes the retrieval of the user's avatar image, and has placeholder code for popping the assignee view controller. Now able to listen to, add, and change task audio in task details, which includes more layer operations (colors are probably off for now). Also included CSAudioPlotViewController changes to allow for a stand-alone mode which will show dismissal / save controls when required to - looks decent enough. functional. Included state cells for °No comments° on a task to mitigate weird empty space problem. Lots of other small tweaks and improvements.

+++ CommSync/CommSync/CSUserRealmModel.m
@@ -84,7 +84,7 @@
 -(NSString *) getPicture
 {
     if(!self.avatar) return @"Avatar-1";
-    return [NSString stringWithFormat:@"Avatar%d", self.avatar];
+    return [NSString stringWithFormat:@"Avatar%ld", (long)self.avatar];
 }
 
 @end
